requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2.1
django-grappelli==3.0.6
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
multidict==6.0.4
Pillow==9.5.0
python-telegram-bot==20.2
requests==2.30.0
rfc3986==1.5.0
sniffio==1.3.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.2
yarl==1.9.2
